backend/requirements.txt
fastapi==0.111.1
uvicorn==0.30.3
sqlalchemy==2.0.31
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
black==24.4.2